ProductService의 조회 메서드 분리
