+++ app/src/main/java/com/codesoom/assignment/application/ToyQueryService.java
@@ -1,0 +1,29 @@
+package com.codesoom.assignment.application;
+
+import com.codesoom.assignment.ProductNotFoundException;
+import com.codesoom.assignment.domain.Product;
+import com.codesoom.assignment.domain.ProductRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+
+@Service
+public class ToyQueryService implements ProductQueryService {
+
+    private ProductRepository productRepository;
+
+    public ToyQueryService(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Override
+    public Product findById(Long id) {
+        return productRepository.findById(id)
+                .orElseThrow(() -> new ProductNotFoundException(id));
+    }
+
+    @Override
+    public Collection<Product> findAll() {
+        return productRepository.findAll();
+    }
+}
